+++ src/layout/component/aside.vue
@@ -146,6 +146,12 @@ watch(
 			if (layoutAsideScrollbarRef.value) layoutAsideScrollbarRef.value.update();
 		}
 		if (layout === 'classic' && isClassicSplitMenu) return false;
+	}
+);
+// 监听用户权限切换，用于演示 `权限管理 -> 前端控制 -> 页面权限` 权限切换不生效
+watch(
+	() => routesList.value,
+	() => {
 		setFilterRoutes();
 	}
 );